Binder JList - JList MitarbeiterProjekte ProjektMitarbeiter View
git-svn-id: file:///home/node/svn/project_se/goran/EPU_Office@179 bde5ac68-c2e4-480a-bd7c-0859b87ef666

+++ src/model/ERechnung.java
@@ -15,7 +15,7 @@ public class ERechnung extends AbstractObject {
     private Kontakt kontakt;
     private String scanPfad;
     private Date datum;
-    private Integer preis;
+    private Double preis;
 
     @Override
     public String toString() {
@@ -41,11 +41,11 @@ public class ERechnung extends AbstractObject {
         this.kontakt = kontakt;
     }
 
-    public int getPreis() {
+    public Double getPreis() {
         return preis;
     }
 
-    public void setPreis(int preis) {
+    public void setPreis(Double preis) {
         this.preis = preis;
     }
